camcontrol: clarify relative position math

+++ include/main/dll/CAM/camcontrol.h
@@ -4,7 +4,7 @@
 #include "ghidra_import.h"
 
 void camcontrol_updateTargetFeedback(void);
-void camcontrol_getRelativePosition(double param_1,int param_2,float *param_3,float *param_4,
+void camcontrol_getRelativePosition(f32 param_1,int param_2,float *param_3,float *param_4,
                                     float *param_5,float *param_6,int param_7);
 void camcontrol_loadTriggeredCamAction(int triggerType,uint actionNo,char triggerMode);
 void *camcontrol_loadCamAction(int actionNo);
